migrating example screen
